% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_complement.R
\name{aa_complement}
\alias{aa_complement}
\title{A Function to complete the aa sequence}
\usage{
aa_complement(required_file, wt_aa, max_sigma = NULL)
}
\arguments{
\item{required_file}{path to "replicates.RData"}

\item{wt_aa}{wt amino acid sequence}

\item{max_sigma}{upper limit of allowable sigma}
}
\description{
This function allow you complete aa sequence in "replicates.RData"
generated by DiMSum, and creat "replicates_fullseq.RData" which
are required for MoCHI.
}
