% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio3D_chain_hetatm_distances.R
\name{bio3D_chain_hetatm_distances}
\alias{bio3D_chain_hetatm_distances}
\title{A Function to get distances from chain to hetatm}
\usage{
bio3D_chain_hetatm_distances(
  input_file,
  chain_query = "A",
  hetatm_target = "GNP"
)
}
\arguments{
\item{input_file}{path to PDB file (required)}

\item{chain_query}{query chain id (default:A)}

\item{hetatm_target}{target hetatm id (default:GNP)}
}
\value{
data.table with hetatm to chain (side-chain) heavy atom distances
}
\description{
A Function to get distances from chain to hetatm
}
