% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio3D_minimum_interchain_distances.R
\name{bio3D_minimum_interchain_distances}
\alias{bio3D_minimum_interchain_distances}
\title{A Function to get minimum interchain distances from PDB}
\usage{
bio3D_minimum_interchain_distances(
  input_file,
  chain_query = "A",
  chain_target = "B",
  residue_query = NULL,
  residue_target = NULL
)
}
\arguments{
\item{input_file}{path to PDB file (required)}

\item{chain_query}{query chain id (default:A)}

\item{chain_target}{target chain id (default:B)}

\item{residue_query}{query residue range (default:all)}

\item{residue_target}{target residue range (default:all)}
}
\value{
data.table with minimum inter-chain (side-chain) heavy atom distances
}
\description{
Calculate minimum inter-chain (side-chain) heavy atom distances.
}
