% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio3D_relative_angle.R
\name{bio3D_relative_angle}
\alias{bio3D_relative_angle}
\title{A Function to get relative angle from PDB}
\usage{
bio3D_relative_angle(input_file, chain = "A")
}
\arguments{
\item{input_file}{path to PDB file (required)}

\item{chain}{chain id (default:A)}
}
\value{
data.table with relative angles for all residues
}
\description{
Get relative angle of c-beta to c-alpha vector w.r.t. center of mass to c-alpha vector from PDB file.
}
