% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio3D_secondary_structure.R
\name{bio3D_secondary_structure}
\alias{bio3D_secondary_structure}
\title{A Function to get secondary structure from PDB}
\usage{
bio3D_secondary_structure(input_file, chain = "A")
}
\arguments{
\item{input_file}{path to PDB file (required)}

\item{chain}{chain id (default:A)}
}
\value{
data.table with secondary structure
}
\description{
Get secondary structure.
}
