% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio3D_temperature_factor.R
\name{bio3D_temperature_factor}
\alias{bio3D_temperature_factor}
\title{A Function to get temperature factor from PDB}
\usage{
bio3D_temperature_factor(input_file, chain = "A")
}
\arguments{
\item{input_file}{path to PDB file (required)}

\item{chain}{chain id (default:A)}
}
\value{
data.table with residue-level temperature factor (B-factor) column values
}
\description{
Get residue-level data from temperature factor (B-factor) columm in PDB file.
}
