% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddG_data_assay.R
\name{ddG_data_assay}
\alias{ddG_data_assay}
\title{A Function to obtain statistical values of interest from ddG data}
\usage{
ddG_data_assay(input, wt_aa)
}
\arguments{
\item{input}{path to MoCHI output ".txt" file}

\item{wt_aa}{wt amino acid sequence}
}
\value{
data table with summarized statistical values
}
\description{
#' @description
This function allows you to get statistical values of interest  from ddG data.
}
