% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddG_heatmap.R
\name{ddG_heatmap}
\alias{ddG_heatmap}
\title{A Function to get ddG heatmap}
\usage{
ddG_heatmap(input, wt_aa, title = "folding free energy change")
}
\arguments{
\item{input}{path to MoCHI output ".txt" file}

\item{wt_aa}{wt amino acid sequence}

\item{title}{heatmap title}
}
\value{
ddG heatmap
}
\description{
#' @description
This function allows you to get ddG heatmap.
}
