% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_heatmap.R
\name{fitness_heatmap}
\alias{fitness_heatmap}
\title{A Function to get fitness heatmap}
\usage{
fitness_heatmap(input, wt_aa, title = "fitness", legend_limits = NULL)
}
\arguments{
\item{input}{single mutant fitness data table}

\item{wt_aa}{wt amino acid sequence}

\item{title}{heatmap title}

\item{legend_limits}{range of legend}
}
\value{
fitness heatmap
}
\description{
This function allows you to get fitness heatmap.
}
