% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nor_fitness.R
\name{nor_fitness}
\alias{nor_fitness}
\title{A Function to normalize fitness from different blocks}
\usage{
nor_fitness(...)
}
\arguments{
\item{...}{paths to "replicate.RData"}
}
\value{
data.table with normalized fitness and sigma
}
\description{
This function allows you to normalize fitness from
different blocks dimsum output data.
}
