% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nor_fitness_single_mut.R
\name{nor_fitness_single_mut}
\alias{nor_fitness_single_mut}
\title{A Function to get single aa mutants fitness data after normalization}
\usage{
nor_fitness_single_mut(input)
}
\arguments{
\item{input}{fitness data table after normalization}
}
\value{
data.table with single aa mutants
}
\description{
This function allows you to get single aa mutants fitness data,
which are required for downstream  heatmap.
}
