% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nor_predict.R
\name{nor_predict}
\alias{nor_predict}
\title{A Function to normalize predicted fitness from MoCHI}
\usage{
nor_predict(pre_file, assay = "Abundance", ...)
}
\arguments{
\item{pre_file}{"predicted_phenotypes_all.txt" from MoCHI}

\item{assay}{assay name}

\item{...}{paths to "replicate.RData"}
}
\value{
data.table with predicted fitness
}
\description{
A Function to normalize predicted fitness from MoCHI
}
