% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D_fold.R
\name{plot2D_fold}
\alias{plot2D_fold}
\title{A Function to plot2D folding fitness against folding ddG}
\usage{
plot2D_fold(input, model, block, RT = 0.001987 * (273 + 30))
}
\arguments{
\item{input}{normalized predicted phenotypes data table}

\item{model}{MoCHI linears_weights model}

\item{block}{target block number}
}
\value{
plot about folding fitness against folding ddG
}
\description{
This function allows you to plot folding fitness against folding ddG.
}
