% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D_vresus.R
\name{plot2D_vresus}
\alias{plot2D_vresus}
\title{A Function to plot2D folding predicted fitness against observed fitness}
\usage{
plot2D_vresus(input, assay, block)
}
\arguments{
\item{input}{normalized predicted phenotypes data table}

\item{assay}{assay name}

\item{block}{target block number}
}
\value{
Nothing
}
\description{
This function allows you to plot folding predicted fitness against observed fitness.
}
