% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D_bind.R
\name{plot3D_bind}
\alias{plot3D_bind}
\title{A Function to plot3D binding fitness against folding ddG and binding ddG}
\usage{
plot3D_bind(
  input,
  model,
  assay,
  block,
  output_file = "./",
  RT = 0.001987 * (273 + 30)
)
}
\arguments{
\item{input}{normalized predicted phenotypes data table}

\item{model}{MoCHI linears_weights model}

\item{assay}{assay name}

\item{block}{target block number}

\item{output_file}{plot output path}
}
\value{
Nothing
}
\description{
This function allows you to plot binding fitness against folding ddG and binding ddG.
}
