% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbg_density_plot.R
\name{rbg_density_plot}
\alias{rbg_density_plot}
\title{A function to draw density plots of false vresus true mutations}
\usage{
rbg_density_plot(input_dir, output_dir = NULL)
}
\arguments{
\item{input_dir}{path to "filter_variants.RData"}

\item{output_dir}{output path (default=input_dir)}
}
\value{
Nothing
}
\description{
A function to draw density plots of false vresus true mutations
}
