% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbg_get_doubles.R
\name{rbg_get_doubles}
\alias{rbg_get_doubles}
\title{A Function to find double mutations from DiMSum "variant_data_merge.tsv"}
\usage{
rbg_get_doubles(input_dt)
}
\arguments{
\item{input_dt}{data table of "variant_data_merge.tsv"}
}
\value{
data table with double mutations
}
\description{
A Function to find double mutations from DiMSum "variant_data_merge.tsv"
}
