% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbg_get_singles.R
\name{rbg_get_singles}
\alias{rbg_get_singles}
\title{A Function to find single mutations from DiMSum "variant_data_merge.tsv"}
\usage{
rbg_get_singles(input_dt)
}
\arguments{
\item{input_dt}{data table of "variant_data_merge.tsv"}
}
\value{
data table with single mutations
}
\description{
A Function to find single mutations from DiMSum "variant_data_merge.tsv"
}
