% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbg_nick_library.R
\name{rbg_nick_library}
\alias{rbg_nick_library}
\title{A Function to obtain mutations distribution from nick library}
\usage{
rbg_nick_library(input_dir_list, output_dir = "./", block = c("1", "2", "3"))
}
\arguments{
\item{input_dir_list}{list contain path to "variant_data_merge.tsv"}

\item{output_dir}{output path (default="./")}

\item{block}{block information of input_dir_list (default=c('1','2','3'))}
}
\value{
Nothing
}
\description{
A Function to obtain mutations distribution from nick library
}
